hart: correctly set `mtval` based on exception code
previously when executing an instruction, the bits of the instruction
were stored in `mtval` on any exception. The RISC-V spec states that if
misaligned/faulty loads/stores write to `mtval` at all it must be with
the faulty virtual address. Fixing this fixes a kernel crash when linux
is probing for hardware misaligned access support as this broken
behaviour led to an access fault in the handler.