move all runnable devices to their own threads
this moves the CLINT, PLIC, and UART, freeing up the main thread for
solely running the hart.

each device has a `task` function to be run in its own thread, as well
as a mutex and condition variable pair for signaling. each device
performs its processing at its own rate/frequency before waiting to run
again, putting itself to sleep instead of busy-waiting if the time till
next run is above some threshold (the CLINT's busy-waiting is also its
own function - maintaining the timer).

each device task thread is woken up on mmio write and read to the
associated device, so as to acquire the mutex lock on the device to
perform any needed internal processing for the non-idempotent memory
access. after the access the lock is restored to the task thread to run
the device and sleep.