plic: do not run in separate thread
running the PLIC in its own thread added unnecessary message-passing
overhead and latency in waking-up the thread to do any processing that
could be done by the device/hart thread interracting with the PLIC
itself.

this makes the threads interracting with the PLIC (a hart through MMIO,
an external device setting or unsetting an interrupt pending bit) assume
the role of the PLIC and updating any external interrupt pending bits on
non-idempotent memory accesses and signalling interrupts which could
affect `mip.meip`. access to the PLIC for this is guarded by a mutex.

this reduces interrupt latency as interrupt raises and claims are
immediately reflected in `mip.meip`, as opposed to waiting for the PLIC
thread to periodically run to update it.

this also reduces CPU consumption by not requiring a separate PLIC thread.