hart: fix wfi implementation
previously the hart would not sleep on a wfi if an interrupt could be
taken, meaning it would sleep for the whole wfi duration if interrupts
were globally disabled. this is incorrect, as the privileged spec states
that the hart can only go to sleep (or remain sleeping) if an enabled
interrupts on the hart are or become pending, regardless of the state of
the global interrupt enable bits `status.mie`, `status.sie`, the
interrupt delegation register `mideleg`. we now check for this condition
instead on entering wfi. during sleep the hart already gets woken up by
any interrupt bit change.